Refactored appointment booking app
User now selects the available time slot from the host, and that time is now added to the users calendar, instead of modifying the hosts calendar.

The user sends the host an invite which the host can choose to do whatever with. It is up to the host to update their calendar so that users can see which appointment times are available.

Users can only book one appointment per week.